+++ MORE/docs/release-notes.md
@@ -5,6 +5,7 @@
 ## Upcoming release
 
 ### Tooth's Emporium of Wonderful Ideas event
+* Changed the Fidget card.
 * Fixed the Fidget card not being available, and the Fiddle card being available twice instead.
 * Fixed the Fiddle card not triggering artifacts' on draw effects.